requirements.txt
asgiref==3.7.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.1
django-allauth==64.0.0
django-crispy-forms==2.2
django-js-asset==2.2.0
django-mptt==0.16.0
django-widget-tweaks==1.5.0
google-auth==2.33.0
idna==3.7
jwt==1.3.1
pillow==10.2.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
rsa==4.9
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.2
